// frontend/src/api/model/healthPlanPracticeRead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */
import type { Nomenclador } from './nomenclador';
import type { PracticeChapter } from './practiceChapter';
import type { PracticeType } from './practiceType';

export interface HealthPlanPracticeRead {
  id: string;
  health_plan_id: string;
  practice_id: string;
  practice_code: string;
  practice_name: string;
  nomenclador: Nomenclador;
  chapter: PracticeChapter;
  practice_type: PracticeType;
  practice_requires_authorization: boolean;
  practice_waiting_period_days: number;
  is_covered: boolean;
  waiting_period_days: number | null;
  effective_waiting_period_days: number;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  copayment_amount: string;
  requires_authorization: boolean;
  notes: string | null;
  created_at: string;
}